+++ benches/iai_benchmark.rs
@@ -9,7 +9,7 @@ use rand::{Rng, SeedableRng};
 pub struct MyElement(usize);
 
 impl PrimeBagElement for MyElement {
-    fn into_prime_index(&self) -> usize {
+    fn to_prime_index(&self) -> usize {
         self.0
     }